refactor: moved default styles to a separate file
