+++ packages/super-editor/src/core/super-converter/v2/importer/docxImporter.js
@@ -1,6 +1,7 @@
 import { getInitialJSON } from '../docxHelper.js';
 import { carbonCopy } from '../../../utilities/carbonCopy.js';
 import { twipsToInches } from '../../helpers.js';
+import { DEFAULT_LINKED_STYLES } from '../../exporter-docx-defs.js';
 import { tableNodeHandlerEntity } from './tableImporter.js';
 import { drawingNodeHandlerEntity } from './imageImporter.js';
 import { trackChangeNodeHandlerEntity } from './trackChangesImporter.js';
@@ -29,661 +30,6 @@ import { getDefaultStyleDefinition } from './paragraphNodeImporter.js';
  * @typedef {{handlerName: string, handler: NodeHandler}} NodeHandlerEntry
  */
 
-export const DEFAULT_LINKED_STYLES = {
-  Normal: {
-    type: 'element',
-    name: 'w:style',
-    attributes: {
-      'w:type': 'paragraph',
-      'w:default': '1',
-      'w:styleId': 'Normal',
-    },
-    elements: [
-      {
-        type: 'element',
-        name: 'w:name',
-        attributes: {
-          'w:val': 'Normal',
-        },
-      },
-      {
-        type: 'element',
-        name: 'w:qFormat',
-      },
-    ],
-  },
-  Title: {
-    type: 'element',
-    name: 'w:style',
-    attributes: {
-      'w:type': 'paragraph',
-      'w:styleId': 'Title',
-    },
-    elements: [
-      {
-        type: 'element',
-        name: 'w:name',
-        attributes: {
-          'w:val': 'Title',
-        },
-      },
-      {
-        type: 'element',
-        name: 'w:basedOn',
-        attributes: {
-          'w:val': 'Normal',
-        },
-      },
-      {
-        type: 'element',
-        name: 'w:next',
-        attributes: {
-          'w:val': 'Normal',
-        },
-      },
-      {
-        type: 'element',
-        name: 'w:link',
-        attributes: {
-          'w:val': 'TitleChar',
-        },
-      },
-      {
-        type: 'element',
-        name: 'w:uiPriority',
-        attributes: {
-          'w:val': '10',
-        },
-      },
-      {
-        type: 'element',
-        name: 'w:qFormat',
-      },
-      {
-        type: 'element',
-        name: 'w:rsid',
-        attributes: {
-          'w:val': '00B72667',
-        },
-      },
-      {
-        type: 'element',
-        name: 'w:pPr',
-        elements: [
-          {
-            type: 'element',
-            name: 'w:spacing',
-            attributes: {
-              'w:after': '80',
-            },
-          },
-          {
-            type: 'element',
-            name: 'w:contextualSpacing',
-          },
-        ],
-      },
-      {
-        type: 'element',
-        name: 'w:rPr',
-        elements: [
-          {
-            type: 'element',
-            name: 'w:rFonts',
-            attributes: {
-              'w:asciiTheme': 'majorHAnsi',
-              'w:eastAsiaTheme': 'majorEastAsia',
-              'w:hAnsiTheme': 'majorHAnsi',
-              'w:cstheme': 'majorBidi',
-            },
-          },
-          {
-            type: 'element',
-            name: 'w:spacing',
-            attributes: {
-              'w:val': '-10',
-            },
-          },
-          {
-            type: 'element',
-            name: 'w:kern',
-            attributes: {
-              'w:val': '28',
-            },
-          },
-          {
-            type: 'element',
-            name: 'w:sz',
-            attributes: {
-              'w:val': '56',
-            },
-          },
-          {
-            type: 'element',
-            name: 'w:szCs',
-            attributes: {
-              'w:val': '56',
-            },
-          },
-        ],
-      },
-    ],
-  },
-  Subtitle: {
-    type: 'element',
-    name: 'w:style',
-    attributes: {
-      'w:type': 'paragraph',
-      'w:styleId': 'Subtitle',
-    },
-    elements: [
-      {
-        type: 'element',
-        name: 'w:name',
-        attributes: {
-          'w:val': 'Subtitle',
-        },
-      },
-      {
-        type: 'element',
-        name: 'w:basedOn',
-        attributes: {
-          'w:val': 'Normal',
-        },
-      },
-      {
-        type: 'element',
-        name: 'w:next',
-        attributes: {
-          'w:val': 'Normal',
-        },
-      },
-      {
-        type: 'element',
-        name: 'w:link',
-        attributes: {
-          'w:val': 'SubtitleChar',
-        },
-      },
-      {
-        type: 'element',
-        name: 'w:uiPriority',
-        attributes: {
-          'w:val': '11',
-        },
-      },
-      {
-        type: 'element',
-        name: 'w:qFormat',
-      },
-      {
-        type: 'element',
-        name: 'w:rsid',
-        attributes: {
-          'w:val': '00B72667',
-        },
-      },
-      {
-        type: 'element',
-        name: 'w:pPr',
-        elements: [
-          {
-            type: 'element',
-            name: 'w:numPr',
-            elements: [
-              {
-                type: 'element',
-                name: 'w:ilvl',
-                attributes: {
-                  'w:val': '1',
-                },
-              },
-            ],
-          },
-          {
-            type: 'element',
-            name: 'w:spacing',
-            attributes: {
-              'w:after': '160',
-            },
-          },
-        ],
-      },
-      {
-        type: 'element',
-        name: 'w:rPr',
-        elements: [
-          {
-            type: 'element',
-            name: 'w:rFonts',
-            attributes: {
-              'w:asciiTheme': 'minorHAnsi',
-              'w:eastAsiaTheme': 'majorEastAsia',
-              'w:hAnsiTheme': 'minorHAnsi',
-              'w:cstheme': 'majorBidi',
-            },
-          },
-          {
-            type: 'element',
-            name: 'w:color',
-            attributes: {
-              'w:val': '595959',
-              'w:themeColor': 'text1',
-              'w:themeTint': 'A6',
-            },
-          },
-          {
-            type: 'element',
-            name: 'w:spacing',
-            attributes: {
-              'w:val': '15',
-            },
-          },
-          {
-            type: 'element',
-            name: 'w:sz',
-            attributes: {
-              'w:val': '28',
-            },
-          },
-          {
-            type: 'element',
-            name: 'w:szCs',
-            attributes: {
-              'w:val': '28',
-            },
-          },
-        ],
-      },
-    ],
-  },
-  Heading1: {
-    type: 'element',
-    name: 'w:style',
-    attributes: {
-      'w:type': 'paragraph',
-      'w:styleId': 'Heading1',
-    },
-    elements: [
-      {
-        type: 'element',
-        name: 'w:name',
-        attributes: {
-          'w:val': 'heading 1',
-        },
-      },
-      {
-        type: 'element',
-        name: 'w:basedOn',
-        attributes: {
-          'w:val': 'Normal',
-        },
-      },
-      {
-        type: 'element',
-        name: 'w:next',
-        attributes: {
-          'w:val': 'Normal',
-        },
-      },
-      {
-        type: 'element',
-        name: 'w:link',
-        attributes: {
-          'w:val': 'Heading1Char',
-        },
-      },
-      {
-        type: 'element',
-        name: 'w:uiPriority',
-        attributes: {
-          'w:val': '9',
-        },
-      },
-      {
-        type: 'element',
-        name: 'w:qFormat',
-      },
-      {
-        type: 'element',
-        name: 'w:rsid',
-        attributes: {
-          'w:val': '00233D7B',
-        },
-      },
-      {
-        type: 'element',
-        name: 'w:pPr',
-        elements: [
-          {
-            type: 'element',
-            name: 'w:keepNext',
-          },
-          {
-            type: 'element',
-            name: 'w:keepLines',
-          },
-          {
-            type: 'element',
-            name: 'w:spacing',
-            attributes: {
-              'w:before': '360',
-              'w:after': '80',
-            },
-          },
-          {
-            type: 'element',
-            name: 'w:outlineLvl',
-            attributes: {
-              'w:val': '0',
-            },
-          },
-        ],
-      },
-      {
-        type: 'element',
-        name: 'w:rPr',
-        elements: [
-          {
-            type: 'element',
-            name: 'w:rFonts',
-            attributes: {
-              'w:asciiTheme': 'majorHAnsi',
-              'w:eastAsiaTheme': 'majorEastAsia',
-              'w:hAnsiTheme': 'majorHAnsi',
-              'w:cstheme': 'majorBidi',
-            },
-          },
-          {
-            type: 'element',
-            name: 'w:color',
-            attributes: {
-              'w:val': '0F4761',
-              'w:themeColor': 'accent1',
-              'w:themeShade': 'BF',
-            },
-          },
-          {
-            type: 'element',
-            name: 'w:sz',
-            attributes: {
-              'w:val': '40',
-            },
-          },
-          {
-            type: 'element',
-            name: 'w:szCs',
-            attributes: {
-              'w:val': '40',
-            },
-          },
-        ],
-      },
-    ],
-  },
-  Heading2: {
-    type: 'element',
-    name: 'w:style',
-    attributes: {
-      'w:type': 'paragraph',
-      'w:styleId': 'Heading2',
-    },
-    elements: [
-      {
-        type: 'element',
-        name: 'w:name',
-        attributes: {
-          'w:val': 'heading 2',
-        },
-      },
-      {
-        type: 'element',
-        name: 'w:basedOn',
-        attributes: {
-          'w:val': 'Normal',
-        },
-      },
-      {
-        type: 'element',
-        name: 'w:next',
-        attributes: {
-          'w:val': 'Normal',
-        },
-      },
-      {
-        type: 'element',
-        name: 'w:link',
-        attributes: {
-          'w:val': 'Heading2Char',
-        },
-      },
-      {
-        type: 'element',
-        name: 'w:uiPriority',
-        attributes: {
-          'w:val': '9',
-        },
-      },
-      {
-        type: 'element',
-        name: 'w:semiHidden',
-      },
-      {
-        type: 'element',
-        name: 'w:unhideWhenUsed',
-      },
-      {
-        type: 'element',
-        name: 'w:qFormat',
-      },
-      {
-        type: 'element',
-        name: 'w:rsid',
-        attributes: {
-          'w:val': '00233D7B',
-        },
-      },
-      {
-        type: 'element',
-        name: 'w:pPr',
-        elements: [
-          {
-            type: 'element',
-            name: 'w:keepNext',
-          },
-          {
-            type: 'element',
-            name: 'w:keepLines',
-          },
-          {
-            type: 'element',
-            name: 'w:spacing',
-            attributes: {
-              'w:before': '160',
-              'w:after': '80',
-            },
-          },
-          {
-            type: 'element',
-            name: 'w:outlineLvl',
-            attributes: {
-              'w:val': '1',
-            },
-          },
-        ],
-      },
-      {
-        type: 'element',
-        name: 'w:rPr',
-        elements: [
-          {
-            type: 'element',
-            name: 'w:rFonts',
-            attributes: {
-              'w:asciiTheme': 'majorHAnsi',
-              'w:eastAsiaTheme': 'majorEastAsia',
-              'w:hAnsiTheme': 'majorHAnsi',
-              'w:cstheme': 'majorBidi',
-            },
-          },
-          {
-            type: 'element',
-            name: 'w:color',
-            attributes: {
-              'w:val': '0F4761',
-              'w:themeColor': 'accent1',
-              'w:themeShade': 'BF',
-            },
-          },
-          {
-            type: 'element',
-            name: 'w:sz',
-            attributes: {
-              'w:val': '32',
-            },
-          },
-          {
-            type: 'element',
-            name: 'w:szCs',
-            attributes: {
-              'w:val': '32',
-            },
-          },
-        ],
-      },
-    ],
-  },
-  Heading3: {
-    type: 'element',
-    name: 'w:style',
-    attributes: {
-      'w:type': 'paragraph',
-      'w:styleId': 'Heading3',
-    },
-    elements: [
-      {
-        type: 'element',
-        name: 'w:name',
-        attributes: {
-          'w:val': 'heading 3',
-        },
-      },
-      {
-        type: 'element',
-        name: 'w:basedOn',
-        attributes: {
-          'w:val': 'Normal',
-        },
-      },
-      {
-        type: 'element',
-        name: 'w:next',
-        attributes: {
-          'w:val': 'Normal',
-        },
-      },
-      {
-        type: 'element',
-        name: 'w:link',
-        attributes: {
-          'w:val': 'Heading3Char',
-        },
-      },
-      {
-        type: 'element',
-        name: 'w:uiPriority',
-        attributes: {
-          'w:val': '9',
-        },
-      },
-      {
-        type: 'element',
-        name: 'w:semiHidden',
-      },
-      {
-        type: 'element',
-        name: 'w:unhideWhenUsed',
-      },
-      {
-        type: 'element',
-        name: 'w:qFormat',
-      },
-      {
-        type: 'element',
-        name: 'w:rsid',
-        attributes: {
-          'w:val': '00233D7B',
-        },
-      },
-      {
-        type: 'element',
-        name: 'w:pPr',
-        elements: [
-          {
-            type: 'element',
-            name: 'w:keepNext',
-          },
-          {
-            type: 'element',
-            name: 'w:keepLines',
-          },
-          {
-            type: 'element',
-            name: 'w:spacing',
-            attributes: {
-              'w:before': '160',
-              'w:after': '80',
-            },
-          },
-          {
-            type: 'element',
-            name: 'w:outlineLvl',
-            attributes: {
-              'w:val': '2',
-            },
-          },
-        ],
-      },
-      {
-        type: 'element',
-        name: 'w:rPr',
-        elements: [
-          {
-            type: 'element',
-            name: 'w:rFonts',
-            attributes: {
-              'w:eastAsiaTheme': 'majorEastAsia',
-              'w:cstheme': 'majorBidi',
-            },
-          },
-          {
-            type: 'element',
-            name: 'w:color',
-            attributes: {
-              'w:val': '0F4761',
-              'w:themeColor': 'accent1',
-              'w:themeShade': 'BF',
-            },
-          },
-          {
-            type: 'element',
-            name: 'w:sz',
-            attributes: {
-              'w:val': '28',
-            },
-          },
-          {
-            type: 'element',
-            name: 'w:szCs',
-            attributes: {
-              'w:val': '28',
-            },
-          },
-        ],
-      },
-    ],
-  },
-};
 /**
  *
  * @param {ParsedDocx} docx
